apps/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
prometheus_client==0.20.0
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
alembic==1.13.2
python-dotenv==1.0.1
tqdm==4.66.5
sentence-transformers==3.0.1
torch==2.4.1
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
joblib==1.4.2
prettytable==3.10.0
pytest==8.3.3
httpx==0.27.2
